+++ tests/Gooi.Tests/MessageHandlerTests.cs
@@ -1,7 +1,6 @@
 using Moq;
 using Skinder.Gooi.Azure;
-using Skinder.Gooi.Core;
-using Skinder.Gooi.Core.Infrastructure;
+using Skinder.Gooi.Contracts.Interfaces.Infrastructure;
 
 namespace Skinder.Gooi.Tests;
 
@@ -10,12 +9,11 @@ public class MessageHandlerTests
     [Fact]
     public async Task MessageHandler_Should_Return0_WithBasicSetup()
     {
-        var writer = new Mock<IOutputWriter>();
-        var connectionProperties = new ConnectionProperties { ConnectionString = "Foo"};
-
-        var messageHandler = new MessageHandler(writer.Object, connectionProperties);
+        var writer = new Mock<IInfrastructureOutputWriter>();
+        
+        var messageHandler = new AzureMessageCommandHandler(writer.Object);
         //
-        var result = await messageHandler.Handle(new MessageCommand
+        var result = await messageHandler.Handle(new AzureQueueMessageCommand()
         {
             Message = "Hello"
         }, new CancellationToken());
